builtin: use isize in memdup

+++ vlib/builtin/builtin.c.v
@@ -603,9 +603,9 @@ pub fn free(ptr voidptr) {
 // memdup dynamically allocates a `sz` bytes block of memory on the heap
 // memdup then copies the contents of `src` into the allocated space and
 // returns a pointer to the newly allocated space.
+
 [unsafe]
-pub fn memdup(src voidptr, sz int) voidptr {
-	// pub fn memdup(src voidptr, sz isize) voidptr {
+pub fn memdup(src voidptr, sz isize) voidptr {
 	$if trace_memdup ? {
 		C.fprintf(C.stderr, c'memdup size: %10d\n', sz)
 	}
@@ -619,8 +619,7 @@ pub fn memdup(src voidptr, sz int) voidptr {
 }
 
 [unsafe]
-pub fn memdup_noscan(src voidptr, sz int) voidptr {
-	// pub fn memdup_noscan(src voidptr, sz isize) voidptr {
+pub fn memdup_noscan(src voidptr, sz isize) voidptr {
 	$if trace_memdup ? {
 		C.fprintf(C.stderr, c'memdup_noscan size: %10d\n', sz)
 	}
